requirements.txt
# Azure AI & Foundry
azure-ai-projects>=2.0.0b1
azure-identity>=1.15.0
openai>=1.30.0
azure-storage-blob>=12.18.0
azure-keyvault-secrets>=4.7.0

# RAG & Vector Stores
chromadb>=0.4.0
sentence-transformers>=2.2.0
qdrant-client>=1.7.0

# Testing & QA
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-xdist>=3.5.0
pytest-timeout>=2.1.0

# API & HTTP
requests>=2.31.0
httpx>=0.25.0
aiohttp>=3.9.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.5.0

# Excel & Reporting
openpyxl>=3.1.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
click>=8.1.0
tqdm>=4.66.0
loguru>=0.7.0
tabulate>=0.9.0

# Azure Evaluation & RAGAS
azure-ai-evaluation>=0.1.0
ragas>=0.1.0
datasets>=2.15.0

# PDF Processing for KB
PyPDF2>=3.0.0
reportlab>=4.0.0  # Optional: For generating sample PDFs